refactor(config): move NodeEnv into src/config
The NodeEnv type (previously called nodeEnvType) is an application
specification, rather than a database specification, since in future
it might be used to control other config (e.g. logging). Hence it makes
more sense for it to live together with app-level config rather than
global types.

Furthermore, NodeEnv should be an enum rather than a string union as
there are cases where we want to specify all the valid values (e.g.
in the convict schema), where Object.values can be used for enums.

The DatabaseConfigType type is also used in src/bootstrap/index and
doesn't seem to need to be a global type, so it was moved into that
file.